Moved Mine placement code over to this interactor
